chore: add 'prepareGitHubUser' fixture for onboarding tests after OAuth
