+++ tests/playwright-utils.ts
@@ -7,8 +7,15 @@ import {
 	sessionKey,
 } from '#app/utils/auth.server.ts'
 import { prisma } from '#app/utils/db.server.ts'
+import { MOCK_CODE_GITHUB_HEADER } from '#app/utils/providers/constants.js'
+import { normalizeEmail } from '#app/utils/providers/provider.js'
 import { authSessionStorage } from '#app/utils/session.server.ts'
 import { createUser } from './db-utils.ts'
+import {
+	type GitHubUser,
+	deleteGitHubUser,
+	insertGitHubUser,
+} from './mocks/github.ts'
 
 export * from './db-utils.ts'
 
@@ -59,6 +66,7 @@ async function getOrInsertUser({
 export const test = base.extend<{
 	insertNewUser(options?: GetOrInsertUserOptions): Promise<User>
 	login(options?: GetOrInsertUserOptions): Promise<User>
+	prepareGitHubUser(): Promise<GitHubUser>
 }>({
 	insertNewUser: async ({}, use) => {
 		let userId: string | undefined = undefined
@@ -93,6 +101,30 @@ export const test = base.extend<{
 			return user
 		})
 		await prisma.user.deleteMany({ where: { id: userId } })
+	},
+	prepareGitHubUser: async ({ page }, use, testInfo) => {
+		await page.route(/\/auth\/github(?!\/callback)/, async (route, request) => {
+			const headers = {
+				...request.headers(),
+				[MOCK_CODE_GITHUB_HEADER]: testInfo.testId,
+			}
+			await route.continue({ headers })
+		})
+
+		let ghUser: GitHubUser | null = null
+		await use(async () => {
+			const newGitHubUser = await insertGitHubUser(testInfo.testId)!
+			ghUser = newGitHubUser
+			return newGitHubUser
+		})
+		
+		const user = await prisma.user.findUniqueOrThrow({
+			select: { id: true, name: true },
+			where: { email: normalizeEmail(ghUser!.primaryEmail) },
+		})
+		await prisma.user.delete({ where: { id: user.id } })
+		await prisma.session.deleteMany({ where: { userId: user.id } })
+		await deleteGitHubUser(ghUser!.primaryEmail)
 	},
 })
 export const { expect } = test
